Add queue id
Upon adding a track to the queue, it will be augmented with a queue id.
This is meant to unambiguously identify an entry and should later be
used to e.g. specify which entry should be removed (instead of just
supporting 'head' and 'tail').

Completely untested as kawa exits right away without any error.